Have `solve` method act like `main`
We sometimes want to be able to pass command line arguments to the
solutions or the test runner.

This means we no longer pass a `std::istream` reference, instead if a
data file is requested using the `--data-dir` command line option we
will redirect `std::cin` to point to the data file if it exists.